backend/requirements.txt
django==5.0
djangorestframework==3.14
django-cors-headers==4.3
python-dotenv==1.0
pytest-django==4.5
pytest==7.4
django-rest-knox==4.2.0
django-ratelimit==3.0.1
gunicorn==21.2
